Relocated functions for consistency and organization
