+++ src/utils/petab.py
@@ -2,16 +2,9 @@ import os
 from typing import Dict, Iterable, Tuple, Sequence, List, Optional
 from dataclasses import dataclass
 
-import numpy as np
 import pandas as pd
 import petab
 import petab.v1.C as C
-import amici
-from libsbml import Model
-
-from src.utils import amici as am
-from src.utils import sbml as sbml
-from src.models import cpe_models as CPE
 
 
 @dataclass
@@ -80,152 +73,6 @@ def define_conditions(init_conditions: Dict[str, Sequence[float]]) -> pd.DataFra
     return conditions.set_index(C.CONDITION_ID)
 
 
-def get_meas_from_amici_sim(
-    rdata: amici.ReturnDataView,
-    observables_df: pd.DataFrame,
-    cond_id: str = "none",
-    obs_sigma: float = 0.00,
-) -> pd.DataFrame:
-
-    meas_dfs = []
-    for obs_id in observables_df.index:
-
-        obs_data = rdata.by_id(obs_id)
-        obs_data = np.array(obs_data) * (1 + obs_sigma * np.random.randn(len(obs_data)))
-        num_pts = len(obs_data)
-
-        obs_meas_df = pd.DataFrame(
-            {
-                C.OBSERVABLE_ID: [obs_id] * num_pts,
-                C.SIMULATION_CONDITION_ID: [cond_id] * num_pts,
-                C.TIME: rdata.ts,
-                C.MEASUREMENT: obs_data,
-            }
-        )
-        meas_dfs.append(obs_meas_df)
-    meas_df = pd.concat(meas_dfs, ignore_index=True)
-
-    return meas_df
-
-
-def define_measurements_amici(
-    amici_model: amici.Model,
-    timepoints: Sequence[float],
-    conditions_df: pd.DataFrame,
-    observables_df: pd.DataFrame,
-    obs_sigma: float = 0.00,
-    meas_sigma: float = 0.005,
-    solver: Optional[amici.Solver] = None,
-    debug_return_rdatas: bool = False,
-) -> pd.DataFrame | Tuple[pd.DataFrame, List[amici.ReturnDataView]]:
-
-    measurement_dfs = []
-    rdatas = []
-
-    for cond_id, row in conditions_df.iterrows():
-        # Extract conditions for this row as a dictionary
-        conditions = row.to_dict()
-
-        # Run the simulation with these conditions
-        rdata = am.run_amici_simulation(
-            amici_model, timepoints, conditions, sigma=meas_sigma, solver=solver
-        )
-        rdatas.append(rdata)
-
-        # Generate measurements from the simulation
-        meas_df = get_meas_from_amici_sim(
-            rdata, observables_df, cond_id=str(cond_id), obs_sigma=obs_sigma
-        )
-        measurement_dfs.append(meas_df)
-
-    measurement_df = pd.concat(measurement_dfs, ignore_index=True)
-
-    if debug_return_rdatas:
-        return measurement_df, rdatas
-    return measurement_df
-
-
-def get_meas_from_cpe_sim(
-    cpe_output: Dict[str, np.ndarray],
-    observables_df: pd.DataFrame,
-    cond_id: str = "none",
-    obs_sigma: float = 0.00,
-) -> pd.DataFrame:
-    # Transform CPE_output to measurements_df
-
-    # This should throw an error if observables_df has anything other
-    # than 'xA' and 'xB' in the C.FORMULA column
-
-    meas_dfs = []
-    for obs_id, row in observables_df.iterrows():
-        observables = row.to_dict()
-        #print(observables)
-
-        obs_formula = observables[C.OBSERVABLE_FORMULA]
-        if obs_formula not in ['xA', 'xB']:
-            raise Exception('Invalid observable.')
-
-        #print(cpe_output)
-        obs_data = cpe_output[obs_formula]
-        obs_data = np.array(obs_data) * (1 + obs_sigma * np.random.randn(len(obs_data)))
-        num_pts = len(obs_data)
-
-        obs_meas_df = pd.DataFrame(
-            {
-                C.OBSERVABLE_ID: [obs_id] * num_pts,
-                C.SIMULATION_CONDITION_ID: [cond_id] * num_pts,
-                C.TIME: cpe_output['X_sol'],
-                C.MEASUREMENT: obs_data,
-            }
-        )
-        meas_dfs.append(obs_meas_df)
-    meas_df = pd.concat(meas_dfs, ignore_index=True)
-
-    return meas_df
-
-
-def define_measurements_cpe(
-    cpe_model: CPE.Model,
-    timepoints: Sequence[float],
-    conditions_df: pd.DataFrame,
-    observables_df: pd.DataFrame,
-    obs_sigma: float = 0.00,
-    meas_sigma: float = 0.005,
-    approach: str = "izu",
-    **kwargs,
-) -> pd.DataFrame:
-    # Loop over all conditions in conditions_df
-
-    # Call CPE.run_CPE_sim, pass in args and kwargs
-
-    # Use get_meas_from_cpe_sim to transform output to measurements_df
-
-    # return measurements_df
-
-    measurement_dfs = []
-    cpe_outputs = []
-
-    for cond_id, row in conditions_df.iterrows():
-        # Extract conditions for this row as a dictionary
-        conditions = row.to_dict()
-
-        # Run the simulation with these conditions
-        cpe_output = CPE.run_CPE_sim(
-            cpe_model, timepoints, conditions, sigma=meas_sigma, **kwargs
-        )
-        # cpe_outputs.append(cpe_output)
-
-        # Generate measurements from the simulation
-        meas_df = get_meas_from_cpe_sim(
-            cpe_output, observables_df, cond_id=str(cond_id), obs_sigma=obs_sigma
-        )
-        measurement_dfs.append(meas_df)
-
-    measurement_df = pd.concat(measurement_dfs, ignore_index=True)
-
-    return measurement_df
-
-
 # Maybe add a petab file directory
 def write_petab_files(
     sbml_model_filepath: str,
@@ -257,7 +104,9 @@ def write_petab_files(
     )
 
     # Define PEtab configuration
-    rel_model_filepath = os.path.join("..", model_filename) if petab_dir_name else model_filename
+    rel_model_filepath = (
+        os.path.join("..", model_filename) if petab_dir_name else model_filename
+    )
     yaml_config = {
         C.FORMAT_VERSION: 1,
         C.PARAMETER_FILE: "parameters.tsv",
